made domain helper class
